docs: rename setting component
